Move Pixel struct to own file
